requirements.txt
torch==1.12.1
torchvision==0.13.1
absl-py==1.2.0
Flask==2.0.2
numpy==1.20.3
Pillow==9.2.0
Werkzeug==2.0.2
